+++ src/run_mcmc.cpp
@@ -90,13 +90,14 @@ Rcpp::List run_mcmc(arma::mat rankings, int nmc,
   bool augpair = (constraints.length() > 0);
   bool any_missing = !arma::is_finite(rankings);
 
-  arma::mat missing_indicator;
-  arma::vec assessor_missing;
+  arma::umat missing_indicator;
+  arma::uvec assessor_missing;
 
   if(any_missing){
-    missing_indicator = rankings;
-    missing_indicator.transform( [](double val) { return (arma::is_finite(val)) ? 0 : 1; } );
-    assessor_missing = arma::conv_to<arma::vec>::from(sum(missing_indicator, 0));
+    // Converting to umat will convert NA to 0
+    missing_indicator = arma::conv_to<arma::umat>::from(rankings);
+    missing_indicator.transform( [](int val) { return (val == 0) ? 1 : 0; } );
+    assessor_missing = arma::conv_to<arma::uvec>::from(arma::sum(missing_indicator, 0));
     initialize_missing_ranks(rankings, missing_indicator, assessor_missing);
   } else {
     missing_indicator.reset();